:bug: FT_73: Fix Developer Options getTable() Error and Fixed refetch issues in Generate CSV Changes Only action
- Added arguments to specify calling of forSync is from re-fetch command
- Added validation to block truncate of sync table if re-fetched

+++ app/Services/CDIS/SyncService.php
@@ -35,10 +35,10 @@ class SyncService
      * @param string $table
      * @return \Illuminate\Http\Response
      */
-    public function forSync($limit = 100, $table = 'all', $broadcast = false, $deleteSyncedDone = DeleteSyncedAction::CONVERT_ALL, $showProgress = false)
+    public function forSync($limit = 100, $table = 'all', $broadcast = false, $deleteSyncedDone = DeleteSyncedAction::CONVERT_ALL, $showProgress = false, $isRefetched = false)
     {
-        return $this->transaction(function () use ($limit, $table, $broadcast, $deleteSyncedDone, $showProgress) {
-            if ($broadcast) {
+        return $this->transaction(function () use ($limit, $table, $broadcast, $deleteSyncedDone, $showProgress, $isRefetched) {
+            if ($broadcast && $deleteSyncedDone === DeleteSyncedAction::CONVERT && ! $isRefetched) {
                 CDISSync::truncate();
             }
